WebHook store configuration refactoring
